+++ changetext/changetext.py
@@ -1,1244 +1,10 @@
-import re
-import traceback
-from collections import OrderedDict
-
-from changetext.common_state import get_state
-from changetext.contextual import corr_contextual
-from changetext.corrector import Corrector
+from changetext.corrector import get_corrector
 from changetext.logging_tools import get_logger, log_exceptions
-from changetext.open_brackets import open_brackets
-from changetext.tag_correction import corr_tags, parse_tags
 from changetext.utf16_codec import utf16_codec
-from changetext.utils import (
-    any_in_tag,
-    custom_parse,
-    to_genitive_case,
-    to_genitive_case_list,
-    to_genitive_case_single_noun,
-    get_gender,
-    get_main_word_gender,
-    inflect_adjective,
-    inflect_as_adjective,
-    inflect_collocation,
-    inflect_noun,
-    is_adjective,
-    make_adjective,
-    parse_as_noun,
-)
-from changetext.whole_phrases import whole_phrases
 
-corrector = Corrector()
 
-replaced_parts = OrderedDict(
-    [
-        ("Ремесленникство", "мастерство"),
-        ("ремесленникство", "мастерство"),
-        ("предложить вашей мастерство", "предложить Вашей Мастеровитости"),
-        ("FIRED_MAT-образующий", ""),
-        ("FIRED_MAT", "обожжённый"),
-        ("TALLOW SOAP_MAT-образующий", "мылообразующий"),
-        ("SOAP_MAT-образующий", "мылообразующий"),
-        ("SOAP_MAT", "мыло"),
-        ("TALLOW", "жир"),
-        ("GLAZE_MAT-образующий", "глазуреобразующий"),
-        ("GLAZE_MAT", "глазурь"),
-        ("CAN_GLAZE", "глазуруемый"),
-        ("FLUX", "флюс"),
-        ("GIPSUM", "гипс"),
-        ("DRINK_MAT", "напиток"),
-        ("FAT", "жир"),
-        ("RENDER_MAT", "вытапливаемый"),
-        # ("PRESS_LIQUID_MAT", ""),
-        # ("HONEYCOMB_PRESS_MAT", ""),
-        (" доверенное л ", " доверенное лицо "),  # Temporary fix for 'hearthperson' cutting
-        ("источника в.", "источника воды."),  # Temporary fix for 'No water source.' cutting
-        ("большой, зазубренный", "большой зазубренный"),
-        ("ремесленник мастерская", "мастерская ремесленника"),
-        ("Ремесленник мастерская", "Мастерская ремесленника"),
-    ]
-)
-
-#############################################################################
-
-animals_female = {
-    "собака",
-    "самка",
-    "крольчиха",
-    "гусыня",
-    "утка",
-    "кошка",
-    "ослица",
-    "кобыла",
-    "корова",
-    "овца",
-    "свинья",
-    "коза",
-    "курица",
-    "свинка",
-    "буйволица",
-    "важенка",
-    "лама",
-    "альпака",
-    "цесарка",
-    "пава",
-    "индейка",
-}
-
-body_parts = {
-    "панцирь",
-    "скелет",
-    "искалеченный труп",
-    "останки",
-    "кость",
-    "кожа",
-    "шёлк",
-    "волокна",
-    "шерсть",
-    "мех",
-    "хвост",
-}
-
-
-re_of_material_item = re.compile(r"^[(+*-«☼р]*(из\s[\w\s\-/]+\b)")
-
-
-@corrector.final_corrector(predicate=lambda text: "пол" not in text and re_of_material_item.search(text))
-@open_brackets
-def corr_of_material_item(text, search_result):
-    """
-    >>> corr_of_material_item("риз алевролита мемориал")
-    '≡алевролитовый мемориал'
-    >>> corr_of_material_item("из алевролита доспешная стойка")
-    'алевролитовая доспешная стойка'
-    >>> corr_of_material_item("(из висмутовой бронзы короткие мечи [3])")
-    '(короткие мечи из висмутовой бронзы [3])'
-    """
-    initial_string = search_result.group(1)
-    words = initial_string.split()
-
-    if len(words) == 2:
-        parse = list(filter(lambda x: {"NOUN", "gent"} in x.tag, custom_parse(words[1])))
-        assert len(parse) == 1
-        replacement_string = parse[0].normal_form
-    elif words[1] == "древесины":
-        # Ultra simple case
-        if "дерева" in words:  # 'из древесины миндального дерева'
-            cut_index = words.index("дерева") + 1
-        elif "пекан" in words:  # 'из древесины ореха пекан'
-            cut_index = words.index("пекан") + 1
-        elif any_in_tag({"NOUN", "gent"}, custom_parse(words[2])):  # 'из древесины яблони'
-            cut_index = 3
-        else:
-            cut_index = -1
-        replacement_string = " ".join(words[cut_index:] + words[:cut_index])
-    elif all(any_in_tag({"ADJF", "gent"}, custom_parse(adj)) for adj in words[1:-1]) and any_in_tag(
-        {"NOUN", "gent"}, custom_parse(words[-1])
-    ):
-        # All words after 'из' except the last word are adjectives in genitive
-        # The last is a noun in genitive
-        material = words[-1]
-        gender = get_gender(material, known_tags={"gent"})
-        parse = list(filter(lambda x: {"NOUN", "gent"} in x.tag, custom_parse(material)))
-        material = parse[0].normal_form
-        adjs = words[1:-1]
-        adjs = [inflect_adjective(adj, gender, case="nomn") for adj in adjs]
-        replacement_string = " ".join(adjs) + " " + material
-    # elif (words[2] not in corr_item_general_except and len(words) > 3 and
-    elif (
-        len(words) > 3
-        and any_in_tag({"gent"}, custom_parse(words[1]))
-        and any_in_tag({"NOUN", "gent"}, custom_parse(words[2]))  # The second word is in genitive
-    ):  # The third word is a noun in genitive
-        # Complex case, eg. "из висмутовой бронзы"
-        of_material = " ".join(words[:3])
-        words = words[3:]
-        if len(words) == 1:
-            first_part = words[0]
-        else:
-            obj = words[-1]
-            gender = get_gender(obj, "NOUN")
-            adjs = (inflect_adjective(adj, gender) or adj for adj in words[:-1])
-            first_part = "{} {}".format(" ".join(adjs), obj)
-        replacement_string = first_part + " " + of_material
-    elif any_in_tag({"NOUN", "gent"}, custom_parse(words[1])) and words[1] != "древесины":
-        # Simple case, eg. "из бронзы"
-        of_material = " ".join(words[:2])
-        words = words[2:]
-        item = words[-1]
-
-        for word in words:
-            if any_in_tag({"NOUN", "nomn"}, custom_parse(word)):
-                item = word
-                break
-
-        if of_material in make_adjective:
-            gender = get_gender(item, {"nomn"})
-
-            if gender is None:
-                for item in reversed(words[:-2]):
-                    gender = get_gender(item)
-                    if gender is not None:
-                        break
-            adjective = make_adjective[of_material]
-            adjective = inflect_adjective(adjective, gender)
-            # If there are another adjectives, ensure that they are in the correct gender:
-            for i, word in enumerate(words):
-                if is_adjective(word) and get_gender(word) != gender:
-                    word = inflect_adjective(word, gender)
-                    words[i] = word
-            replacement_string = adjective + " " + " ".join(words)
-        else:
-            replacement_string = " ".join(words) + " " + of_material
-    else:
-        raise ValueError("Unknown case: {!r}".format(text))
-
-    text = text.replace(initial_string, replacement_string)
-    return text
-
-
-@corrector.final_corrector(
-    regex=r"(\(?)(.+)\s(\bяйцо|требуха|железы|железа|мясо|кровь|сукровица|кольца|серьги|амулеты"
-    r"|браслеты|скипетры|короны|статуэтки\b)"
-)
-def corr_item_3(text, search_result):
-    """
-    >>> corr_item_3('рогатый филин яйцо')
-    'яйцо рогатого филина'
-    """
-    group_3 = search_result.group(3)
-    if group_3 in replaced_parts:
-        group_3 = replaced_parts[group_3]
-
-    if search_result.group(2) in make_adjective:
-        group_2 = make_adjective[search_result.group(2)]
-    else:
-        group_2 = to_genitive_case(search_result.group(2))
-
-    text = text.replace(search_result.group(0), search_result.group(1) + group_3 + " " + group_2)
-
-    return text
-
-
-@corrector.final_corrector(regex=r"(\bЛужа|Брызги|Пятно)\s(.+)\s(кровь\b)")
-def corr_blood_stain(_text, search_result):
-    """
-    >>> corr_blood_stain("Лужа кот кровь")
-    'Лужа крови кота'
-    """
-    text = search_result.group(1) + " " + to_genitive_case(search_result.group(3) + " " + search_result.group(2))
-    return text.capitalize()
-
-
-@corrector.final_corrector(regex=r"[\^\W]((приготовленные|рубленная)\s(.+)\s(\w+))")
-def corr_prepared(text, search_result):
-    """
-    >>> corr_prepared(" приготовленные гигантский крот лёгкие")
-    ' приготовленные лёгкие гигантского крота'
-    >>> corr_prepared("(приготовленные северный олень почки [5])")
-    '(приготовленные почки северного оленя [5])'
-    >>> corr_prepared("(приготовленные панда лёгкие [5])")
-    '(приготовленные лёгкие панды [5])'
-    >>> corr_prepared("(рубленная гризли печень [5])")
-    '(рубленная печень гризли [5])'
-    """
-    # TODO:
-    # >>> corr_prepared("(приготовленный волк мозг [5])")
-    # '(приготовленный мозг волка [5])'
-    # >>> corr_prepared(" рубленная гигантский крот лёгкие")
-    # ' рубленные лёгкие гигантского крота'
-    groups = search_result.groups()
-    result = text.replace(groups[0], "{} {} {}".format(groups[1], groups[3], to_genitive_case(groups[2])))
-    return result
-
-
-@corrector.final_corrector(regex=r"(\(?)(.+)\s(из кожи|из шерсти)")
-def corr_item_skin(text, search_result):
-    """
-    >>> corr_item_skin("горный козёл из кожи")
-    'кожа горного козла'
-
-    >>> corr_item_skin("альпака из шерсти")
-    'шерсть альпака'
-    """
-    material = inflect_noun(search_result.group(3).split()[-1], "nomn")  # кожа, шерсть и т.д.
-    text = text.replace(
-        search_result.group(0), search_result.group(1) + material + " " + to_genitive_case(search_result.group(2))
-    )
-    return text
-
-
-@corrector.final_corrector(
-    regex=r"^[Xx\(+*-«☼]*((.+)\s"
-    r"(из волокон"
-    r"|из шёлка"
-    r"|из шерсти"
-    r"|из кожи"
-    r"|из копыт"
-    r"|из кости"
-    r"|из рога"
-    r"|из рогов"
-    r"|из бивней"
-    r"|из панциря"
-    r"|из зубов)"
-    r"\s(\w+\s?\w+))"
-)
-@open_brackets
-def corr_clothes(text, search_result):
-    """
-    >>> corr_clothes("свинохвост из волокон ткань")
-    'ткань из волокон свинохвоста'
-    >>> corr_clothes("(-«пещерный паук из шёлка левый варежка»-)")
-    '(-«левая варежка из шёлка пещерного паука»-)'
-    >>> corr_clothes("(гигантский пещерный паук из шёлка шапка)")
-    '(шапка из шёлка гигантского пещерного паука)'
-    """
-    text = text.replace(
-        search_result.group(1),
-        search_result.group(4) + " " + search_result.group(3) + " " + to_genitive_case(search_result.group(2)),
-    )
-    text = text.replace("левый", "левая")
-    text = text.replace("правый", "правая")
-    return text
-
-
-@corrector.final_corrector(regex=r"(древесина)\s(\w+)\s(брёвна)")
-def corr_wooden_logs(text, search_result):
-    """
-    >>> corr_wooden_logs('древесина дуба брёвна')
-    'дубовые брёвна'
-    """
-    of_wood = "из " + search_result.group(2)
-    if of_wood in make_adjective:
-        adj = inflect_adjective(make_adjective[of_wood], "plur")
-        text = text.replace(search_result.group(0), adj + " " + search_result.group(3))  # берёзовые брёвна
-    else:
-        text = text.replace(
-            search_result.group(0), search_result.group(1) + " " + search_result.group(2)
-        )  # древесина акации
-    return text
-
-
-@corrector.final_corrector(regex=r"((бриолетовый|большой|огранённый|грубый)\s[\w\s-]+)")
-def corr_gem_cutting(text, search_result):
-    """
-    >>> corr_gem_cutting("(бриолетовый восковые опалы)")
-    '(бриолетовые восковые опалы)'
-    >>> corr_gem_cutting("большой шерлы")
-    'большие шерлы'
-    """
-    words = search_result.group(1).split()
-    if words[-1] in body_parts:
-        return corr_item_body_parts(text)
-
-    gender = get_gender(words[-1], {"NOUN", "nomn"})
-
-    new_list = []
-    for word in words[:-1]:
-        if word in make_adjective:
-            adj = make_adjective[word]
-            word = inflect_adjective(adj, gender)
-        elif is_adjective(word):
-            word = inflect_adjective(word, gender)
-        new_list.append(word)
-
-    new_list.append(words[-1])
-
-    return text.replace(search_result.group(0), " ".join(new_list))
-
-
-@corrector.final_corrector(
-    regex=r"(шипованный|огромный|большой|заточенный|гигантский|большой зазубренный)\s(из\s[\w\s]+\b)"
-)
-def corr_weapon_trap_parts(text, search_result):
-    """
-    >>> corr_weapon_trap_parts('гигантский из меди лезвия топоров')
-    'гигантские медные лезвия топоров'
-    >>> corr_weapon_trap_parts('большой зазубренный из берёзы диски')
-    'большие зазубренные берёзовые диски'
-    """
-    adj = search_result.group(1)
-    words = search_result.group(2).split()
-    if " ".join(words[:2]) in make_adjective:
-        material = " ".join(words[:2])
-        obj = " ".join(words[2:])
-        gender = get_main_word_gender(obj)
-        new_adj = inflect_as_adjective(adj, gender)
-        new_word_2 = inflect_adjective(make_adjective[material], gender)
-        text = text.replace(search_result.group(0), "{} {} {}".format(new_adj, new_word_2, obj))
-    else:
-        material = " ".join(words[:3])
-        obj = " ".join(words[3:])
-        gender = get_main_word_gender(obj)
-        assert gender is not None
-        new_adj = inflect_as_adjective(adj, gender)
-        text = text.replace(search_result.group(0), "{} {} {}".format(new_adj, obj, material))
-    return text
-
-
-animal_genders = {"собака": ("пёс", "собака"), "кошка": ("кот", "кошка"), "лошадь": ("конь", "лошадь")}
-
-
-@corrector.preliminary_corrector(regex=r"(\w+), ([♂♀])")
-def corr_animal_gender(text, search_result):
-    gender = "♂♀".index(search_result.group(2))
-    animal = search_result.group(1)
-    if animal not in animal_genders:
-        return None
-    else:
-        return text.replace(search_result.group(0), animal_genders[animal][gender] + ", " + search_result.group(2))
-
-
-@corrector.final_corrector(regex=r"(охотничий|боевой|сырой) (\w+)(\(Ручной\))?")
-def corr_animal(text):
-    text = text.replace("сырой", "сырая")
-    if any(item in text for item in animals_female):
-        text = text.replace("(Ручной)", "(Ручная)")
-        text = text.replace("боевой", "боевая")
-        text = text.replace("Ничей", "Ничья")
-        text = text.replace("охотничий", "охотничья")
-
-    return text
-
-
-possessive_adjectives = {"жаба": "жабий", "корова": "коровий", "медведь": "медвежий"}
-
-replace_containment = {
-    "Семя": "семена",
-    "Специи": "специй",
-    "Самоцвет": "самоцветы",
-    "Слиток/Блок": "слитков/блоков",
-}
-
-materials = {"волокон", "шёлка", "шерсти", "кожи"}
-
-
-@corrector.final_corrector(regex=r"(\b.+)\s(бочка|мешок|ящик)\s\((.*?)(\)|$)")
-@open_brackets
-def corr_container(text, search_result):
-    """
-    >>> corr_container('(дварфийское пиво бочка (из ольхи))')
-    '(Бочка дварфийского пива (ольховая))'
-    >>> corr_container("(дварфийское вино бочка (из клёна) <#8>)")
-    '(Бочка дварфийского вина (кленовая) <#8>)'
-    >>> corr_container("(Семя бочка (из лумбанга) <#10>)")
-    '(Бочка семян (лумбанговая) <#10>)'
-    """
-    initial_string = search_result.group(0)
-    containment = search_result.group(1)
-    if containment in replace_containment:
-        containment = replace_containment[containment]
-    if containment.endswith("кровь"):
-        words = containment.split()
-        if words[0] in possessive_adjectives:
-            words[0] = possessive_adjectives[words[0]]
-            words = to_genitive_case_list(words)
-        else:
-            words = [to_genitive_case_single_noun(words[-1])] + list(to_genitive_case_list(words[:-1]))
-        containment = " ".join(words)
-    elif containment.startswith("из "):
-        containment = containment[3:]  # Words after 'из' are already in genitive case
-    elif containment in {"слитков/блоков", "специй"}:
-        pass  # Already in genitive case
-    elif containment.startswith("семена"):
-        words = containment.split()
-        words[0] = to_genitive_case(words[0])
-        containment = " ".join(words)
-    else:
-        containment = to_genitive_case(containment)
-    container = search_result.group(2)
-    of_material = search_result.group(3)
-    if not of_material:
-        replacement_string = container + " " + containment
-    elif (
-        " " not in of_material
-        and is_adjective(of_material)
-        or of_material in make_adjective
-        or of_material[3:] in make_adjective
-    ):
-        if " " not in of_material and is_adjective(of_material):
-            adjective = of_material
-        elif of_material in make_adjective:
-            adjective = make_adjective[of_material]
-        elif of_material[3:] in make_adjective:
-            adjective = make_adjective[of_material[3:]]
-        else:
-            adjective = None
-        gender = get_gender(container, {"nomn"})
-        adjective = inflect_adjective(adjective, gender)
-        replacement_string = "{} {} ({})".format(container, containment, adjective)
-    else:
-        words = of_material.split()
-        material = None
-        if of_material.startswith("из ") or len(of_material) <= 2:
-            material = of_material
-        elif (
-            len(words) >= 2
-            and words[-2] == "из"
-            and (words[-1] in materials or any(mat.startswith(words[-1]) for mat in materials))
-        ):
-            # Try to fix truncated materail names, eg. '(ямный краситель мешок (гигантский пещерный паук из шёл'
-            if words[-1] not in materials:  # Fix partial material name eg. 'шерст', 'шёлк'
-                candidates = [mat for mat in materials if mat.startswith(words[-1])]
-                if len(candidates) == 1:
-                    words[-1] = candidates[0]
-                else:
-                    material = of_material  # Partial name is not recognized (too short)
-
-            if not material:
-                material = " ".join(words[-2:] + list(to_genitive_case_list(words[:-2])))
-        else:
-            gen_case = list(to_genitive_case_list(of_material.split()))
-            if None not in gen_case:
-                material = "из " + " ".join(gen_case)
-            else:
-                material = of_material
-        replacement_string = "{} {} ({}".format(container, containment, material)
-        if initial_string[-1] == ")":
-            replacement_string += ")"
-    text = text.replace(initial_string, replacement_string.capitalize())
-    return text
-
-
-@corrector.final_corrector(
-    regex=r"(.+)\s(Подъем|Стена|Кластер|валун|склон|Пол Пещеры|лестница вверх/вниз|пол пещеры|"
-    r"Лестница Вверх|Лестница Вниз|галька|деревце|лестница вверх|лестница вниз|подъем|пол)\b"
-)
-def corr_relief(_, search_result):
-    group1 = search_result.group(1)
-    obj = search_result.group(2)
-    if obj == "деревце":
-        if group1.split(" ")[0] == "Мёртвый":
-            text = "Мёртвое деревце ({})".format("".join(search_result.group(0).split(" ")[1:-1]))
-        else:
-            text = "Деревце ({})".format(group1)
-        return text.capitalize()
-
-    if " " in group1:
-        words = group1.split(" ")
-        first_words = []
-        gender = get_main_word_gender(obj)
-
-        for word in words:
-            if word in {"Заснеженный", "Неотесанный", "Влажный"}:
-                if gender is not None:
-                    new_word = inflect_adjective(word, gender)
-                    if not new_word:
-                        new_word = word
-                else:
-                    new_word = word
-                first_words.append(new_word)
-            else:
-                break
-
-        words = words[len(first_words) :]
-
-        if words[0] == "из":
-            words = words[1:]
-        else:
-            words = to_genitive_case_list(words)
-
-        if not first_words:
-            text = "{} из {}".format(obj, " ".join(words))
-        else:
-            text = "{} {} из {}".format(" ".join(first_words), obj, " ".join(words))
-    else:
-        material = group1
-        text = "{} из {}".format(obj, to_genitive_case(material))
-
-    if "иза" in text:
-        text = text.replace(" иза", "")
-    return text.capitalize()
-
-
-@corrector.final_corrector(regex=re.compile(r"\b(Густой|Редкий|Заснеженный)\s(.+)"))
-def corr_adjective_relief(text, search_result):
-    """
-    >>> corr_adjective_relief("Заснеженный Густой овсяница")
-    'Заснеженная густая овсяница'
-    """
-    adjective = search_result.group(1)
-    obj = search_result.group(2)
-
-    if " " in obj:
-        words = obj.split(" ")
-        if is_adjective(words[0]):
-            gender = get_gender(words[-1])
-            new_word = inflect_adjective(words[0], gender, "nomn")
-            text = text.replace(words[0], new_word)
-            new_word = inflect_adjective(adjective, gender, "nomn")
-            text = text.replace(adjective, new_word)
-    else:
-        gender = get_gender(obj)
-        new_word = inflect_adjective(adjective, gender, "nomn")
-        if new_word:
-            text = new_word + " " + obj
-
-    return text.capitalize()
-
-
-@corrector.final_corrector(
-    regex=r"^{?((\w+\s?\w+?|)\s" r"(панцирь|скелет|труп|останки|кость|кожа|шёлк|волокна|шерсть|мех|хвост|голень))}?\b"
-)
-def corr_item_body_parts(text, search_result):
-    initial_string = search_result.group(1)
-    words = search_result.group(2).split()
-    if words[-1] in {"частичный", "искалеченный"}:
-        replacement_string = "{} {} {}".format(
-            words[-1],
-            search_result.group(3),
-            " ".join(to_genitive_case_list(words[:-1])),
-        )
-    else:
-        if any("GRND" in custom_parse(word)[0].tag for word in words):  # Ignore participles
-            return None
-        replacement_string = search_result.group(3) + " " + " ".join(to_genitive_case_list(words))
-    return text.replace(initial_string, replacement_string.capitalize())
-
-
-@corrector.final_corrector(regex=r"\b(Делать|Изготовить)\s([\w\s]*)(стекло|хрусталь)([\w\s]*)")
-def corr_craft_glass(text, search_result):  # TODO: Combine into single crafting-related function
-    material = search_result.group(3)
-    material_gender = get_gender(material)
-    words = search_result.group(2).split()
-    product = search_result.group(4).split()
-    verb = search_result.group(1)
-    if not product:
-        verb = "Варить"
-        adjectives = (inflect_adjective(adj, material_gender, "accs", animated=False) for adj in words)
-        result = "{} {} {}".format(verb, " ".join(adjectives), material)
-    else:
-        index = next(
-            (i for i, item in enumerate(words) if item in {"грубое", "зелёное", "прозрачное", "грубый"}), len(words)
-        )
-        product_adjectives = words[:index]
-        if any_in_tag({"NOUN", "nomn"}, custom_parse(product[0])):
-            product_gender = get_gender(product[0])
-            product[0] = inflect_noun(product[0], case="accs")
-        else:
-            product_gender = get_gender(product[-1])
-            product_adjectives += product[:-1]
-            product = [inflect_noun(product[-1], case="accs")]
-
-        product_adjectives = [
-            inflect_adjective(adj, product_gender, case="accs", animated=False) for adj in product_adjectives
-        ]
-        material_adjectives = [
-            inflect_adjective(adj, material_gender, case="gent", animated=False) for adj in words[index:]
-        ]
-
-        material = inflect_noun(material, case="gent")
-        product_words = product_adjectives + product
-        material_words = material_adjectives + [material]
-        result = "{} {} из {}".format(verb, " ".join(product_words), " ".join(material_words))
-
-    return text.replace(search_result.group(0), result)
-
-
-@corrector.final_corrector(regex=r"(Делать|Изготовить|Украшать)([\w\s/]+)$")
-def corr_craft_general(text, search_result):
-    verb = search_result.group(1)
-    words = search_result.group(2).split()
-    product = None
-    if len(words) > 1:
-        for i, word in enumerate(words[1:], 1):
-            if any_in_tag({"NOUN", "nomn"}, custom_parse(word)) and word not in make_adjective:
-                product = " ".join(words[i:])
-                words = words[:i]
-                break
-    else:
-        product = words[0]
-        words = []
-
-    product_gender = get_main_word_gender(product)
-
-    if " " not in product:
-        orig_form = {"plur" if product_gender == "plur" else "sing", "inan"}
-        product = inflect_noun(product, "accs", orig_form=orig_form)
-        assert product is not None
-    else:
-        product = inflect_collocation(product, {"accs"})
-
-    if words:
-        if len(words) == 1 and words[0] not in make_adjective and not is_adjective(words[0]):
-            material = inflect_noun(words[0], "gent", orig_form={"nomn", "inan"})  # рог -> (из) рога
-            assert material is not None
-            result = "{} {} из {}".format(verb, product, material)
-        else:
-            adjectives = [
-                make_adjective[word] if word in make_adjective else word if is_adjective(word) else None
-                for word in words
-            ]
-            assert all(adj is not None for adj in adjectives)
-            adjectives = [inflect_adjective(adj, product_gender, "accs", animated=False) for adj in adjectives]
-            result = "{} {} {}".format(verb, " ".join(adjectives), product)
-    else:
-        result = "{} {}".format(verb, product)
-
-    return text.replace(search_result.group(0), result).capitalize()
-
-
-@corrector.final_corrector(regex=r"(^Ковать|^Делать|^Чеканить|^Изготовить|^Кузница)\s(из\s[\w\s?]+\b)")
-def corr_forge(_, search_result):
-    verb = search_result.group(1)
-    words = search_result.group(2).split()
-    assert len(words) >= 3
-    if any_in_tag({"ADJF", "gent"}, custom_parse(words[1])) and any_in_tag(  # The second word is an adj in gent
-        {"NOUN", "gent"}, custom_parse(words[2])
-    ):  # The third word is a noun in gent
-        of_material = words[:3]
-        obj = words[3:]
-    else:
-        assert any_in_tag({"NOUN", "gent"}, custom_parse(words[1]))
-        of_material = words[:2]
-        obj = words[2:]
-
-    of_material = " ".join(of_material)
-    item_index = None
-    if len(obj) == 1:
-        item_index = 0
-        parse = custom_parse(obj[item_index])
-        p = parse_as_noun(parse)
-        gender = get_gender(obj[item_index], known_tags={"nomn"})
-        if not any_in_tag({"accs"}, p):
-            obj[0] = p[0].inflect({"accs"}).word
-    else:
-        parse = None
-        gender = None
-        for i, word in enumerate(obj):
-            parse = custom_parse(word)
-            p = parse_as_noun(parse)
-            if p:
-                item_index = i
-                gender = get_gender(obj[item_index])
-                obj[i] = p[0].inflect({"accs"}).word
-                break  # Words after the 'item' must be leaved in genitive case
-            elif not any_in_tag("accs", parse):
-                obj[i] = parse[0].inflect({"accs"}).word
-
-    if not any_in_tag("accs", parse):
-        obj[item_index] = parse[0].inflect({"accs"}).word
-
-    if verb == "Кузница":
-        verb = "Ковать"
-
-    if of_material in make_adjective:
-        material = inflect_adjective(make_adjective[of_material], gender, "accs", animated=False)
-        text = verb + " " + material + " " + " ".join(obj)
-    else:
-        text = verb + " " + " ".join(obj) + " " + of_material
-
-    return text.capitalize()
-
-
-@corrector.final_corrector(
-    regex=r"^(Инкрустировать Готовые товары с"
-    r"|Инкрустировать Предметы обстановки с"
-    r"|Инкрустировать Снаряды с"
-    r"|Огранить)\s(.+)"
-)
-def corr_jewelers_shop(_, search_result):
-    first_part = search_result.group(1)
-    words = search_result.group(2).split()
-    if first_part == "Огранить":
-        # accusative case
-        tags = None
-        if words[0] == "из":
-            words = words[1:]
-            tags = {"gent"}
-        item = words[-1]
-        gender = get_gender(item, known_tags=tags)
-        words = [inflect_adjective(word, gender, "accs", animated=False) for word in words[:-1]]
-        parse = list(filter(lambda x: {gender, "inan"} in x.tag, custom_parse(item)))
-        if item == "адамантина":
-            item = "адамантин"
-        else:
-            item = parse[0].inflect({"accs"}).word
-        words.append(item)
-    else:
-        # instrumental/ablative case ('incrust with smth')
-        words = [custom_parse(word)[0].inflect({"ablt"}).word for word in words if word != "из"]
-
-    if first_part.endswith(" с"):
-        first_part = first_part[:-2]
-    text = first_part + " " + " ".join(words)
-    return text.capitalize()
-
-
-@corrector.final_corrector(
-    regex=r"(.*)\s(лесное убежище|крепость|селение|горный город|городок|гробница|пригорки)\s(.+)"
-)
-def corr_settlement(_, search_result):
-    adjective = search_result.group(1).strip()
-    settlement = search_result.group(2)
-    name = search_result.group(3)
-
-    if len(adjective) == 0:
-        return "{} {}".format(settlement.capitalize(), name.capitalize())
-
-    if adjective in {"Покинуть", "Разрушить"}:
-        return
-
-    gender = get_main_word_gender(settlement)
-
-    if " " not in adjective:
-        adjective_2 = inflect_adjective(adjective, gender)
-    else:
-        adjective_2 = " ".join(inflect_adjective(word, gender) for word in adjective.split(" "))
-
-    if adjective_2 is None:
-        adjective_2 = adjective
-
-    return "{} {} {}".format(adjective_2.capitalize(), settlement, name.capitalize())
-
-
-# re_material_selection = re.compile(r"(металл|кожа|пряжа|растительное волокно|дерево|шёлк)\s(.+)")
-#
-# reagents = {
-#     # реагенты
-#     "сырой рыба": "свежая рыба",
-# }
-#
-#
-# # выбор материала
-# def corr_material_selection(text):
-#     search_result = re_material_selection.search(text)
-#     if search_result.group(2) in reagents:
-#         new_word = reagents[search_result.group(2)]
-#     else:
-#         new_word = search_result.group(2)
-#     if search_result.group(1) == "пряжа" or search_result.group(1) == "растительное волокно":
-#         material = make_adjective[search_result.group(1)]
-#         text = new_word + " " + material
-#         return text.capitalize()
-#     gender = get_gender(new_word)
-#     material = make_adjective[search_result.group(1)][gender]
-#     text = material + " " + new_word
-#     return text.capitalize()
-
-
-@corrector.final_corrector(regex=r"(.+)\s(кожа|кость|волокно|шёлк)\b")
-def corr_animal_material(_, search_result):
-    return search_result.group(2) + " " + to_genitive_case(search_result.group(1))
-
-
-@corrector.final_corrector(regex=r"(\w+) приостановили строительство (.*)\.")
-def corr_stopped_construction(_, search_result):
-    """
-    >>> corr_stopped_construction(" дварфы приостановили строительство Стена.")
-    'Дварфы приостановили строительство стены.'
-    >>> corr_stopped_construction(" дварфы приостановили строительство Ремесленник мастерская.")
-    'Дварфы приостановили строительство мастерской ремесленника.'
-    >>> corr_stopped_construction(" дварфы приостановили строительство Ювелирная мастерская.")
-    'Дварфы приостановили строительство ювелирной мастерской.'
-    """
-    subj = search_result.group(1)
-    obj = search_result.group(2)
-
-    if "Ремесленник мастерская" in obj:
-        gen_case_obj = " ".join(
-            to_genitive_case(word) for word in reversed(obj.split())
-        )  # Put words into genitive case separately
-    else:
-        gen_case_obj = to_genitive_case(obj)
-
-    return ("{} приостановили строительство {}.".format(subj, gen_case_obj)).capitalize()
-
-
-# Корректировка для окончания s - перевод существительного во множественное число или глагола в 3-е лицо ед.ч.
-re_ending_s = re.compile(r"(\d+\s)?([а-яёА-ЯЁ][а-яёА-ЯЁ\s]*)e?s\b")
-
-
-def corr_ending_s_internal(text):
-    parse = [x for x in custom_parse(text) if {"NOUN", "nomn", "sing"} in x.tag or {"VERB", "2per"} in x.tag]
-
-    if not parse:
-        # Cannot determine part of speech
-        return None
-
-    new_forms = set()
-
-    if parse[0].tag.POS == "NOUN":
-        new_forms.add(parse[0].inflect({"plur"}).word)
-    else:  # parse[0].tag.POS == 'VERB':
-        new_forms.add(parse[0].inflect({"3per", "sing"}).word)
-
-    if len(new_forms) > 1:
-        # Cannot determine part of speech because of ambiguity
-        return None
-
-    return new_forms.pop()
-
-
-dict_ending_s = {
-    "готовая еда": "готовая еда",
-    "питьё": "питьё",
-    "стул": "стулья",
-    "доспешная стойка": "доспешные стойки",
-    "оружейная стойка": "оружейные стойки",
-    "дублёная шкура": "дублёные шкуры",
-    "большой самоцвет": "большие самоцветы",
-    "баклер": "баклеры",
-    "оружие": "оружие",
-    "крышка люка": "крышки люка",
-    "ручная мельница": "ручные мельницы",
-    "ловушка для животных": "ловушки для животных",
-    "часть ловушки": "части ловушек",
-    "музыкальный инструмент": "музыкальные инструменты",
-    "наконечник стрелы баллисты": "наконечники стрелы баллисты",
-    "часть тела": "части тела",
-    "конечность/тело гипс": "гипс для конечностей тела",
-    "Элитный борец": "Элитные борцы",
-    "Лорд топора": "Лорды топора",
-    "Лорд булавы": "Лорды булавы",
-    "Лорд молота": "Лорды молота",
-    "Мастер меча": "Мастера меча",
-    "Мастер копья": "Мастера копья",
-}
-
-
-def corr_ending_s(text):
-    search_result = re_ending_s.search(text)
-    number = search_result.group(1)
-    group2 = search_result.group(2)
-    if number and " " not in group2:
-        number = int(number)
-        parse = [x for x in custom_parse(group2) if {"NOUN", "nomn", "sing"} in x.tag]
-        assert len(parse) == 1
-        replacement_string = "{:d} {}".format(number, parse[0].make_agree_with_number(number).word)
-    elif group2 in dict_ending_s:
-        replacement_string = dict_ending_s[group2]
-    elif " " not in group2:
-        new_form = corr_ending_s_internal(group2)
-        if new_form:
-            replacement_string = new_form
-        else:
-            return None
-    else:
-        words = group2.split()
-        if words[-1] in dict_ending_s:
-            words[-1] = dict_ending_s[words[-1]]
-            replacement_string = " ".join(words)
-        else:
-            new_form = corr_ending_s_internal(words[-1])
-            if new_form:
-                words[-1] = new_form
-                replacement_string = " ".join(words)
-            else:
-                return None
-
-    return text.replace(search_result.group(0), replacement_string)
-
-
-@corrector.preliminary_corrector(regex=re_ending_s)
-def corr_ending_s_loop(text, _search_result):
-    result = text
-    while re_ending_s.search(text):
-        s1 = corr_ending_s(text)
-        if s1 is None:
-            break
-        text = s1
-        result = text
-    return result
-
-
-# Clothier's shop
-cloth_subst = {
-    "ткань": ("Шить", "из ткани"),
-    "шёлк": ("Шить", "шёлковый"),
-    "пряжа": ("Вязать", "из пряжи"),
-    "кожа": ("Шить", "из кожи"),
-}
-
-
-@corrector.final_corrector(regex=r"(Делать|Изготовить|Вышивать|Ткать) (ткань|шёлк|пряжа|кожа)(\s?\w*)")
-def corr_clothiers_shop(_, search_result):
-    verb = search_result.group(1)
-    material = search_result.group(2)
-    product = search_result.group(3).strip()
-
-    if not product:
-        return None  # Leave as is eg. 'Ткать шёлк'
-    elif verb == "Вышивать":
-        parse = custom_parse(material)[0]
-        if material == "пряжа":
-            verb = "Вязать"
-            material = parse.inflect({"gent"}).word
-            return "{} {} из {}".format(verb, product, material)
-        else:
-            material = parse.inflect({"loct"}).word
-            return "{} {} на {}".format(verb, product, material)
-    else:
-        if product in {"щит", "баклер"}:
-            of_material = cloth_subst[material][1]  # Leave 'Делать'/'Изготовить' verb
-        else:
-            verb, of_material = cloth_subst[material]
-
-        if product == "верёвка":
-            verb = "Вить"
-
-        gender = get_gender(product, {"nomn"})
-        if gender == "femn":
-            product_accus = inflect_noun(product, case="accs")
-        else:
-            product_accus = product
-
-        if material in make_adjective:
-            material_adj = inflect_adjective(make_adjective[material], gender, "accs", animated=False)
-            return " ".join([verb, material_adj, product_accus])
-        else:
-            return " ".join([verb, product_accus, of_material])
-
-
-@corrector.preliminary_corrector(regex=r"были(\w+)")
-def corr_werebeast(text, search_result):
-    """
-    >>> corr_werebeast("Ura Wuspinicen, былимуравьед")
-    'Ura Wuspinicen, муравьед-оборотень'
-    >>> corr_werebeast(" былимуравьед крепко держится!")
-    ' муравьед-оборотень крепко держится!'
-    """
-    return text.replace(search_result.group(0), search_result.group(1) + "-оборотень")
-
-
-@corrector.final_corrector(regex=r"(.+)\s(стал)\s(.+)\.")
-def corr_become(_, search_result):
-    """
-    >>> corr_become("Udil Vuthiltobul стал рекрут.")
-    'Udil Vuthiltobul стал рекрутом.'
-    >>> corr_become("Udil Vuthiltobul стал рыбник.")
-    'Udil Vuthiltobul стал рыбником.'
-    >>> corr_become("`***CLOTH' Rovodokir стал портной.")
-    "`***CLOTH' Rovodokir стал портным."
-    >>> corr_become("Животное вырос и стал Ничей козёл.")
-    'Животное выросло и стало ничьим козлом.'
-    >>> corr_become("Животное вырос и стал Ничей важенка.")
-    'Животное выросло и стало ничьей важенкой.'
-    """
-    subj = search_result.group(1)
-    verb = search_result.group(2)
-    words = search_result.group(3)
-    words = inflect_collocation(words, {"ablt"})
-    if subj.startswith("Животное"):
-        return "Животное выросло и стало {}.".format(words)
-    else:
-        return "{} {} {}.".format(subj, verb, words)
-
-
-@corrector.preliminary_corrector(regex=r"с (его|её|ваш) ([\w\s]*)")
-def corr_with_his(text, search_result):
-    """
-    >>> corr_with_his("ладонь с его левое предплечье!")
-    'ладонь своим левым предплечьем!'
-    >>> corr_with_his("ладонь с ваш левое предплечье!")
-    'ладонь своим левым предплечьем!'
-    """
-    return text.replace(
-        search_result.group(0), "своим {}".format(inflect_collocation(search_result.group(2), {"ablt"}))
-    )
-
-
-@corrector.final_corrector(regex=r"([\w\s]+) (кольцо|кольца)")
-def corr_rings(text, search_result):
-    obj = search_result.group(2)
-    description = search_result.group(1)
-    return text.replace(search_result.group(0), "{} из {}".format(obj, to_genitive_case(description)))
-
-
-# Title eg. "Histories of Greed and Avarice" for the Linux version
-histories_adjs = {
-    "Greed": " жадности",
-    "Avarice": "б алчности",
-    "Jealousy": " зависти",
-    "Cupidity": " скупости",
-    "Gluttony": "б обжорстве",
-    "Industry": "производстве",
-    "Enterprise": "предприимчивости",
-    "Resourcefulness": "находчивости",
-    "Determination": "решительности",
-    "Mettle": "отваге",
-    "Dynamism": "стремительности",
-    "Labor": "работе",
-    "Toil": "труде",
-    "Diligence": "усердии",
-    "Exertion": "напряжении",
-    "Tenacity": "стойкости",
-    "Perseverance": "упорстве",
-}
-
-
-@corrector.final_corrector(regex=r"Histories of (\w+) and (\w+)")
-def corr_histories_of(_, search_result):
-    return "Истории о{} и {}".format(histories_adjs[search_result.group(1)], histories_adjs[search_result.group(2)])
-
-
-@corrector.preliminary_corrector(predicate=lambda text: "Я " in text and "колодец" in text)
-def corr_well(text, _):
-    """
-    >>> corr_well('Я колодец')  # I am well
-    'Я в порядке'
-
-    >>> corr_well('Я чувствую колодец')  # I am feeling well
-    'Я чувствую себя хорошо'
-    """
-    text = text.replace("колодец", "хорошо")
-    if "чувствую" in text and "себя" not in text:
-        text = text.replace("чувствую", "чувствую себя")
-    elif "делаю хорошо" in text:
-        text = text.replace("делаю хорошо", "в порядке")
-    elif "был хорошо" in text:
-        text = text.replace("был хорошо", "в порядке")
-    elif "хорошо" in text:
-        text = text.replace("хорошо", "в порядке")
-    return text
-
-
-@corrector.preliminary_corrector(predicate=lambda text: "рублены" in text and "рубленый " not in text)
-def corr_minced(text, _):
-    s1 = ""
-    while "рублены" in text and "рубленый" not in text:
-        x, _, text = text.partition("рублены")
-        s1 += x + "рубленый "
-
-    return s1 + text
-
-
-@corrector.final_corrector(predicate=lambda text: text.startswith("Вы нашли из "))
-def corr_you_struck(text, _):
-    """
-    >>> corr_you_struck('Вы нашли из пиролюзита!')
-    'Вы нашли пиролюзит!'
-    >>> corr_you_struck('Вы нашли из каменного угля!')
-    'Вы нашли каменный уголь!'
-    """
-    text = text.rstrip("!")
-    you_struck, of, material = text.partition(" из ")
-
-    words = material.split()
-    assert len(words) >= 1
-    if len(words) == 1:
-        result = inflect_noun(words[0], "accs")
-    else:
-        result = inflect_collocation(material, {"accs"})
-
-    return you_struck + " " + result + "!"
-
-
-@corrector.preliminary_corrector(regex=re.compile(r"(он|она|вы)\s+(не\s+)?(имеете?)", flags=re.IGNORECASE))
-def corr_someone_has(text, search_result):
-    pronoun = search_result.group(1).lower()
-    if pronoun == "он":
-        replacement_string = "у него"
-    elif pronoun == "она":
-        replacement_string = "у неё"
-    elif pronoun == "вы":
-        replacement_string = "у вас"
-    else:
-        return text
-
-    if search_result.group(0)[0].isupper():
-        replacement_string = replacement_string.capitalize()
-
-    if search_result.group(2):
-        replacement_string += " нет"
-
-    text = text.replace(search_result.group(0), replacement_string)
-    assert isinstance(text, str), text
-    return text
-
-
-@corrector.preliminary_corrector(regex=r"(имеете?|был)\s+(\w+)")
-def corr_has_verb(text, search_result):
-    """
-    >>> corr_has_verb(" имеет создал ")
-    ' создал '
-    >>> corr_has_verb(" был создал ")
-    ' создал '
-    >>> corr_has_verb(" был создать ")
-    ' создал '
-    >>> corr_has_verb(" имеет пришёл ")
-    ' пришёл '
-    >>> corr_has_verb(" имеет упал ")
-    ' упал '
-    >>> corr_has_verb(" имеет стрямкал ")
-    ' стрямкал '
-    """
-    if search_result:
-        word = search_result.group(2)
-        parse = [p for p in custom_parse(word) if p.tag.POS in ("VERB", "INFN")]
-        if parse:
-            if not any({"past"} in p.tag for p in parse):
-                word = parse[0].inflect({"VERB", "past", "sing"}).word
-            return text.replace(search_result.group(0), word)
-
-
-@corrector.preliminary_corrector(regex=r"[a-z](в <[a-z]+>)")
-def corr_in_ending(text, search_result):
-    return text.replace(search_result.group(1), "in")
-
-
-@corrector.preliminary_corrector(predicate=lambda text: text in whole_phrases)
-def corr_whole_phrases(text, _):
-    return whole_phrases[text]
-
-
-@corrector.preliminary_corrector(regex=r"цвет ([\w\s]*)цвета")
-def corr_color_of_color(text, search_result):
-    """
-    >>> corr_color_of_color("цвет серебристого цвета")
-    'серебристый цвет'
-    >>> corr_color_of_color("цвет цвета морской волны")
-    'цвет морской волны'
-    """
-    if search_result:
-        if not search_result.group(1):
-            replacement = "цвет"
-        else:
-            color = search_result.group(1).strip()
-            replacement = "{} цвет".format(inflect_collocation(color, {"nomn", "masc"}))
-        return text.replace(search_result.group(0), replacement)
-
-
-@corrector.preliminary_corrector(predicate=lambda text: any(item in text for item in replaced_parts))
-def corr_replace_parts(text, _):
-    result = text
-
-    for item in replaced_parts:
-        if item in text:
-            text = text.replace(item, replaced_parts[item])
-            result = text
-
-    return result
-
-
-corrector.preliminary_corrector()(lambda text, _: corr_contextual(text))
-
-
-@corrector.preliminary_corrector(predicate=lambda _: get_state().prev_tail)
-def add_prev_tail(text, prev_tail):
-    text = prev_tail + text
-    get_state().prev_tail = ""
-    return text
-
-
-@corrector.preliminary_corrector(
-    predicate=lambda text: "<" in text and ">" in text and "<нет " not in text and "<#" not in text
-)
-def corr_tags_outer(text, _):
-    try:
-        result = corr_tags(text)
-    except (AssertionError, ValueError) as err:
-        print("corr_tags() raises exception {!r}:".format(err))
-        print(traceback.format_exc())
-        result = " ".join(
-            part.strip(" ") if not part.startswith("<") else part.strip("<>").partition(":")[2]
-            for part in parse_tags(text)
-        )
-    return result
-
-
-############################################################################
 def change_text(text):
-    return corrector.change_text(text)
+    return get_corrector().change_text(text)
 
 
 @utf16_codec
