+++ changetext/replaced_parts.py
@@ -1,0 +1,30 @@
+from collections import OrderedDict
+
+replaced_parts = OrderedDict(
+    [
+        ("Ремесленникство", "мастерство"),
+        ("ремесленникство", "мастерство"),
+        ("предложить вашей мастерство", "предложить Вашей Мастеровитости"),
+        ("FIRED_MAT-образующий", ""),
+        ("FIRED_MAT", "обожжённый"),
+        ("TALLOW SOAP_MAT-образующий", "мылообразующий"),
+        ("SOAP_MAT-образующий", "мылообразующий"),
+        ("SOAP_MAT", "мыло"),
+        ("TALLOW", "жир"),
+        ("GLAZE_MAT-образующий", "глазуреобразующий"),
+        ("GLAZE_MAT", "глазурь"),
+        ("CAN_GLAZE", "глазуруемый"),
+        ("FLUX", "флюс"),
+        ("GIPSUM", "гипс"),
+        ("DRINK_MAT", "напиток"),
+        ("FAT", "жир"),
+        ("RENDER_MAT", "вытапливаемый"),
+        # ("PRESS_LIQUID_MAT", ""),
+        # ("HONEYCOMB_PRESS_MAT", ""),
+        (" доверенное л ", " доверенное лицо "),  # Temporary fix for 'hearthperson' cutting
+        ("источника в.", "источника воды."),  # Temporary fix for 'No water source.' cutting
+        ("большой, зазубренный", "большой зазубренный"),
+        ("ремесленник мастерская", "мастерская ремесленника"),
+        ("Ремесленник мастерская", "Мастерская ремесленника"),
+    ]
+)
